use package extensions
