+++ Project.toml
@@ -15,6 +15,7 @@ SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 [compat]
 Espresso = "0.6"
 InteractiveUtils = "<0.0.1, 1.4"
+LazySets = "5"
 LinearAlgebra = "<0.0.1, 1.4"
 MacroTools = "0.5"
 MultivariatePolynomials = "0.3, 0.4, 0.5"
@@ -22,3 +23,13 @@ RecipesBase = "0.6, 0.7, 0.8, 1.0, 1.1, 1.2"
 Requires = "1.3.1"
 SparseArrays = "<0.0.1, 1.4"
 julia = "1.6"
+
+[extensions]
+LazySetsExt = "LazySets"
+
+[extras]
+LazySets = "b4f0291d-fe17-52bc-9479-3d1a343d9043"
+
+[weakdeps]
+LazySets = "b4f0291d-fe17-52bc-9479-3d1a343d9043"
+Requires = "ae029012-a4dd-5104-9daa-d747884805df"
